journal entry apis updated

+++ src/main/java/com/mayur/journalApp/service/JournalEntryService.java
@@ -41,15 +41,28 @@ public class JournalEntryService {
     public Optional<JournalEntry> findById(ObjectId id){
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(String userName,ObjectId id){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveUser(user);
-        journalEntryRepository.deleteById(id);
+    public boolean deleteById(String userName,ObjectId id){
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+
+        }
+        catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured");
+        }
+
+        return removed;
+
     }
-    public void deleteAll() {
-        journalEntryRepository.deleteAll();
-    }
+
+
+
 
 }
 // controller->service->repository
